requirements.txt
Flask
jira
jproperties
lib
wheel
pandas
requests
urllib3
astroid
bencode.py
dill
docopt
et-xmlfile
isort
lazy-object-proxy
mccabe
observable
openpyxl
pipreqs
platformdirs
pylint
setuptools
tomli
wrapt
xmlutils
yarg
sslyze
cffi